requirements.txt
django
djangorestframework
requests
psycopg
django-rest-swagger